I renamed entryModel to db.js and the change affected addentry,getAll,modifyentry,specificentry. even I created new file called connect for connecting with database
